Fix implementation issues, update tests

+++ src/main/kotlin/com/github/alexeysol/geekregimeapiusers/exceptions/UserNotFoundException.kt
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 class UserNotFoundException(
     message: String? = Constants.USER_NOT_FOUND,
-    userId: Int? = null
+    userId: Long? = null
 ) : BaseApiUsersException(
     message = message,
     invalidKeyValuePair = userId?.let{ Pair("id", userId.toString()) }
