+++ src/main/kotlin/com/github/alexeysol/geekregimeapiusers/repositories/CredentialsRepository.kt
@@ -1,8 +1,8 @@
 package com.github.alexeysol.geekregimeapiusers.repositories
 
-import com.github.alexeysol.geekregimeapiusers.entities.Credentials
+import com.github.alexeysol.geekregimeapiusers.models.entities.Credentials
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CredentialsRepository : CrudRepository<Credentials, Int>
+interface CredentialsRepository : CrudRepository<Credentials, Long>
